requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==23.0.0
httpie==3.2.3
idna==3.7
inflection==0.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
packaging==24.1
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uuid==1.30
